requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.8.0rc2
Django==5.1.2
django-extensions==3.2.3
django-otp==1.5.4
django-phonenumber-field==8.0.0
frozenlist==1.4.1
idna==3.10
MarkupSafe==3.0.2
multidict==6.1.0
oauthlib==3.2.2
phonenumbers==8.13.47
pillow==11.0.0
propcache==0.2.0
psycopg==3.2.3
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-decouple==3.8
python3-openid==3.2.0
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
twilio==9.3.4
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
yarl==1.16.0
